requirements.txt
pyspark==4.1.1
python-dotenv==1.2.1
aiokafka==0.13.0